+++ src/main/java/com/example/techbridge/domain/member/service/MemberQueryService.java
@@ -1,0 +1,50 @@
+package com.example.techbridge.domain.member.service;
+
+import com.example.techbridge.domain.member.dto.MemberDetailResponse;
+import com.example.techbridge.domain.member.entity.Member;
+import com.example.techbridge.domain.member.entity.Member.Role;
+import com.example.techbridge.domain.member.exception.MemberNotFoundException;
+import com.example.techbridge.domain.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class MemberQueryService {
+
+    private final MemberRepository memberRepository;
+
+    public Member findById(Long id) {
+        return memberRepository.findById(id)
+            .orElseThrow(MemberNotFoundException::new);
+    }
+
+    public Member findByUsername(String username) {
+        return memberRepository.findByUsername(username)
+            .orElseThrow(MemberNotFoundException::new);
+    }
+
+    public Member findByEmail(String email) {
+        return memberRepository.findByEmail(email)
+            .orElseThrow(MemberNotFoundException::new);
+    }
+
+    public boolean existsByUsername(String username) {
+        return memberRepository.existsByUsername(username);
+    }
+
+    public boolean existsByEmail(String email) {
+        return memberRepository.existsByEmail(email);
+    }
+
+    public Page<MemberDetailResponse> findAll(Role role, Pageable pageable) {
+        Page<Member> page = (role == null) ? memberRepository.findAll(pageable)
+            : memberRepository.findByRole(role, pageable);
+
+        return page.map(MemberDetailResponse::of);
+    }
+}
